fix: Namespaced linetype
